refactor: move app from _util to cli.main
