+++ weasel/__init__.py
@@ -1,11 +1,1 @@
-from ._util import app
-
-# These are the actual functions, NOT the wrapped CLI commands. The CLI commands
-# are registered automatically and won't have to be imported here.
-from .cli.assets import project_assets
-from .cli.clone import project_clone
-from .cli.document import project_document
-from .cli.dvc import project_update_dvc
-from .cli.pull import project_pull
-from .cli.push import project_push
-from .cli.run import project_run
+from .cli import app
