+++ weasel/__main__.py
@@ -1,3 +1,3 @@
-from ._util import COMMAND, app
+from .cli.main import COMMAND, app
 
 app(prog_name=COMMAND)
