+++ weasel/cli/assets.py
@@ -8,10 +8,10 @@ import requests
 import typer
 from wasabi import msg
 
-from .._util import PROJECT_FILE, Arg, Opt, SimpleFrozenDict, app, download_file
-from .._util import get_checksum, get_git_version, git_checkout, load_project_config
-from .._util import parse_config_overrides
+from .._util import SimpleFrozenDict, download_file, get_checksum, get_git_version
+from .._util import git_checkout, load_project_config, parse_config_overrides
 from ..util import ensure_path, working_dir
+from .main import PROJECT_FILE, Arg, Opt, app
 
 # Whether assets are extra if `extra` is not set.
 EXTRA_DEFAULT = False
