+++ weasel/cli/clone.py
@@ -6,9 +6,9 @@ from typing import Optional
 from wasabi import msg
 
 from .. import about
-from .._util import COMMAND, PROJECT_FILE, Arg, Opt, app, get_git_version, git_checkout
-from .._util import git_repo_branch_exists
+from .._util import get_git_version, git_checkout, git_repo_branch_exists
 from ..util import ensure_path
+from .main import COMMAND, PROJECT_FILE, Arg, Opt, app
 
 DEFAULT_REPO = about.__projects__
 DEFAULT_PROJECTS_BRANCH = about.__projects_branch__
